get e2e test of feature_demo workflow running

+++ tests/context_managers/celery_workers_manager.py
@@ -52,7 +52,8 @@ class CeleryWorkersManager:
 
     def __init__(self, app: Celery):
         self.app = app
-        self.running_workers = []
+        self.running_workers = set()
+        self.run_worker_processes = set()
         self.worker_processes = {}
         self.echo_processes = {}
 
@@ -145,7 +146,7 @@ class CeleryWorkersManager:
         """
         self.app.worker_main(worker_launch_cmd)
 
-    def launch_worker(self, worker_name: str, queues: List[str], concurrency: int = 1):
+    def launch_worker(self, worker_name: str, queues: List[str], concurrency: int = 1, prefetch: int = 1):
         """
         Launch a single worker. We'll add the process that the worker is running in to the list of worker processes.
         We'll also create an echo process to simulate a celery worker command that will show up with 'ps ux'.
@@ -170,6 +171,8 @@ class CeleryWorkersManager:
             ",".join(queues),
             "--concurrency",
             str(concurrency),
+            "--prefetch-multiplier",
+            str(prefetch),
             f"--logfile={worker_name}.log",
             "--loglevel=DEBUG",
         ]
@@ -187,7 +190,7 @@ class CeleryWorkersManager:
         worker_process = multiprocessing.Process(target=self.start_worker, args=(worker_launch_cmd,))
         worker_process.start()
         self.worker_processes[worker_name] = worker_process
-        self.running_workers.append(worker_name)
+        self.running_workers.add(worker_name)
 
         # Wait for the worker to launch properly
         try:
@@ -206,6 +209,24 @@ class CeleryWorkersManager:
         """
         for worker_name, worker_settings in worker_info.items():
             self.launch_worker(worker_name, worker_settings["queues"], worker_settings["concurrency"])
+
+    def add_run_workers_process(self, pid: int):
+        """
+        Add a process ID for a `merlin run-workers` process to the
+        set that tracks all `merlin run-workers` processes that are
+        currently running.
+        
+        Warning:
+            The process that's added here must utilize the
+            `start_new_session=True` setting of subprocess.Popen. This
+            is necessary for us to be able to terminate all the workers
+            that are started with it safely since they will be seen as
+            child processes of the `merlin run-workers` process.
+
+        Args:
+            pid: The process ID running `merlin run-workers`.
+        """
+        self.run_worker_processes.add(pid)
 
     def stop_worker(self, worker_name: str):
         """
@@ -226,12 +247,16 @@ class CeleryWorkersManager:
                 self.worker_processes[worker_name].kill()
 
         # Terminate the echo process and its sleep inf subprocess
-        os.killpg(os.getpgid(self.echo_processes[worker_name]), signal.SIGTERM)
-        sleep(2)
+        if self.echo_processes[worker_name] is not None:
+            os.killpg(os.getpgid(self.echo_processes[worker_name]), signal.SIGTERM)
+            sleep(2)
 
     def stop_all_workers(self):
         """
         Stop all of the running workers and the processes associated with them.
         """
+        for run_worker_pid in self.run_worker_processes:
+            os.killpg(os.getpgid(run_worker_pid), signal.SIGTERM)
+
         for worker_name in self.running_workers:
             self.stop_worker(worker_name)
